requirements.txt
altgraph==0.17.5
packaging==25.0
pulsectl==24.12.0
Pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2025.11
PyQt6==6.10.2
PyQt6-Charts==6.10.0
PyQt6-Charts-Qt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
pyusb==1.3.1
QtPy==2.4.3
setuptools==80.9.0
superqt==0.7.7
typing_extensions==4.15.0
